Fix the setting of the T value
Hit upon this problem

    $ ppb 2766040588
    inf TiB

This was due to T being set to 0 and then doing a divide by 0.

The reason T was set to 0 was due to the way C does multiplication of
different sized entities. In this case we had a uint32_t * uint16_t,
according to "Section 6.2.1.5 - Usual Arithmetic Conversions of the ANSI
C standard" in this case we'd hit this

    "Otherwise, if either operand has type unsigned int, the other operand
     is converted to unsigned int"

thus the calculation will be done as a 32bit unsigned integer and that
won't hold the answer and 1024*1024*1024*1024 when assigned to an
unsigned int wraps to 0.

The solution is to simply cast one of the mutlipliers to a uint64_t,
which makes us hit an earlier condition

    "If either operand has type unsigned long int, the other operand is
     converted to unsigned long int."

and we now get the right answer

    $ ppb 2766040588
    2.58 GiB

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>